requirements.txt
asgiref==3.2.10
autopep8==1.5.4
cffi==1.14.2
cryptography==3.1
dj-database-url==0.5.0
Django==3.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
